Refactor SubjectController for code consistency

+++ gradecalculator/src/main/java/com/harshit/gradecalculator/controller/SubjectController.java
@@ -51,8 +51,9 @@ public class SubjectController {
 
     // 3. DELETE a Subject
     @DeleteMapping("/delete")
-    public String deleteSubject(@RequestParam Long id) {
+    public String deleteSubject(@RequestParam Integer id) {
         subjectRepository.deleteById(id);
         return "Subject Deleted!";
     }
 }
+
